+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # 1.3.8
 - fix styles without `frontmatter.image`
 - adjust the header text of the list page
+- optimizing the tags styles
 
 ## 1.3.7
 - optimizing styles